[코드최적화]:중복제거
